pm: Split interfaces up into individual files
